+++ go/cron/cleanup_chi_database_cache.go
